==> notebooks/04_feature_engineering.ipynb <==
import pandas as pd
import sqlite3
import matplotlib.pyplot as plt
import seaborn as sns

conn = sqlite3.connect("../churn.db")
df = pd.read_sql("SELECT * FROM customers", conn)

df.head()

df["ChurnBinary"] = df["Churn"].map({"Yes": 1, "No": 0})


df = df.drop(columns=["customerID", "Churn"])


df_encoded = pd.get_dummies(df, drop_first=True)


from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()

numeric_cols = ["tenure", "MonthlyCharges", "TotalCharges"]

df_encoded[numeric_cols] = scaler.fit_transform(df_encoded[numeric_cols])


### Train Test Split

from sklearn.model_selection import train_test_split

X = df_encoded.drop(columns=["ChurnBinary"])
y = df_encoded["ChurnBinary"]

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, stratify=y, random_state=42
)


#save to the file
import os
import pickle

os.makedirs("../data/processed", exist_ok=True)

with open("../data/processed/X_train.pkl", "wb") as f:
    pickle.dump(X_train, f)

with open("../data/processed/X_test.pkl", "wb") as f:
    pickle.dump(X_test, f)

with open("../data/processed/y_train.pkl", "wb") as f:
    pickle.dump(y_train, f)

with open("../data/processed/y_test.pkl", "wb") as f:
    pickle.dump(y_test, f)

print("Train/test splits saved successfully.")
